+++ src/common/types.rs
@@ -1,1 +1,1 @@
-pub type NetworkParams = (Vec<Vec<f64>>, Vec<Vec<f64>>, Vec<Vec<f64>>, Vec<Vec<f64>>);
+pub type NetworkParams = (Vec<Vec<f32>>, Vec<Vec<f32>>, Vec<Vec<f32>>, Vec<Vec<f32>>);
